Update canBeEnabled check to check cache signature as well
